extract/requirements.txt
pytrends
newsapi-python
psycopg2-binary
python-dotenv